+++ ft_printf_str.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf_str.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: uboumedj <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: uboumedj <uboumedj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/05 15:24:01 by uboumedj          #+#    #+#             */
-/*   Updated: 2017/12/12 10:18:26 by uboumedj         ###   ########.fr       */
+/*   Updated: 2017/12/15 04:36:57 by uboumedj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_do_str(const char *str, int *len)
+void	ft_do_str(const char *str, size_t *len)
 {
 	ft_putstr(str);
 	*len += ft_strlen(str);
